requirements.txt
slack_sdk>=3.33.0
python-dotenv>=1.0.0
rich>=13.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
anthropic>=0.79.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
